refactor: IDE 편의성을 위해 index.tsx 파일명을 컴포넌트명으로 변경
- 모든 컴포넌트 파일명을 의미 있는 이름으로 변경
- 배럴 파일 및 import 경로 업데이트
- IDE 탭 구분 및 파일 검색 개선
