+++ src/components/index.ts
@@ -1,3 +1,3 @@
 export { NaverMap } from './Maps'
-export { default as CustomMarker } from './CustomMarker'
-export { default as RestSpotMarker } from './RestSpotMarker'
+export { default as CustomMarker } from './CustomMarker/CustomMarker'
+export { default as RestSpotMarker } from './RestSpotMarker/RestSpotMarker'
