+++ src/pages/NaverPage/components/Logos/Logos.tsx
