+++ src/pages/NaverPage/components/Notice/Notice.tsx
