+++ src/pages/NaverPage/components/PathInfo/PathInfo.tsx
