+++ src/pages/NaverPage/components/RestAreaInfo/RestAreaInfo.tsx
