+++ src/pages/NaverPage/components/Survey/Survey.tsx
