+++ src/pages/NaverPage/components/index.ts
@@ -1,9 +1,9 @@
-export { default as RestAreaInfo } from './RestAreaInfo'
-export { default as InputSubmit } from './InputSubmit'
-export { default as Logos } from './Logos'
-export { default as PathInfo } from './PathInfo'
-export { default as Title } from './Title'
-export { default as RecentSearch } from './RecentSearch'
-export { default as Loading } from './Loading'
-export { default as Survey } from './Survey'
-export { default as Main } from './Main'
+export { default as RestAreaInfo } from './RestAreaInfo/RestAreaInfo'
+export { default as InputSubmit } from './InputSubmit/InputSubmit'
+export { default as Logos } from './Logos/Logos'
+export { default as PathInfo } from './PathInfo/PathInfo'
+export { default as Title } from './Title/Title'
+export { default as RecentSearch } from './RecentSearch/RecentSearch'
+export { default as Loading } from './Loading/Loading'
+export { default as Survey } from './Survey/Survey'
+export { default as Main } from './Main/Main'
